+++ tasks/universal-turing-machine/src/direction.rs
@@ -1,0 +1,5 @@
+pub enum Direction {
+    Left,
+    Right,
+    Stay,
+}
