+++ tasks/universal-turing-machine/src/lib.rs
@@ -1,0 +1,7 @@
+mod direction;
+mod rule;
+mod tm;
+
+pub use direction::*;
+pub use rule::*;
+pub use tm::*;
